+++ lib/Epub/Epub/ParsedText.h
@@ -29,6 +29,7 @@ class ParsedText {
   bool hyphenationEnabled;
   bool bionicReadingEnabled;
   bool guideReadingEnabled;
+  bool boundedTokenGrowth;
   uint8_t wordSpacing;
   BlockStyle blockStyle;
   bool hasRtlWord;
@@ -85,14 +86,15 @@ class ParsedText {
 
  public:
   explicit ParsedText(const bool extraParagraphSpacing, const bool forceParagraphIndents = false,
-                      const bool hyphenationEnabled = false, const bool bionicReadingEnabled = false,
-                      const bool guideReadingEnabled = false, const uint8_t wordSpacing = 0,
-                      const BlockStyle& blockStyle = BlockStyle())
+                       const bool hyphenationEnabled = false, const bool bionicReadingEnabled = false,
+                       const bool guideReadingEnabled = false, const uint8_t wordSpacing = 0,
+                       const BlockStyle& blockStyle = BlockStyle(), const bool boundedTokenGrowth = false)
       : extraParagraphSpacing(extraParagraphSpacing),
         forceParagraphIndents(forceParagraphIndents),
         hyphenationEnabled(hyphenationEnabled),
         bionicReadingEnabled(bionicReadingEnabled),
         guideReadingEnabled(guideReadingEnabled),
+        boundedTokenGrowth(boundedTokenGrowth),
         wordSpacing(wordSpacing),
         blockStyle(blockStyle),
         hasRtlWord(false) {}